rename staffEmailHandler to caseEmailHandler
